Created Redux store and slices also useSelector hook to access store data
