+++ src/index.js
@@ -4,10 +4,18 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
+//redux imports
+import { store } from './store'
+import { Provider } from 'react-redux'
+
+
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    {/* similar to context api wrap our app within provider so that the store is available globally */}
+    <Provider store={store}>
+      <App />
+    </Provider>
   </React.StrictMode>
 );
